MaintainOperatorValueHelper -> ContractFacade integration test
